+++ ui/apps/platform/src/utils/networkGraphUtils.test.constants.js
@@ -1,4 +1,4 @@
-import { filterModes } from 'Containers/Network/Graph/filterModes';
+import { filterModes } from 'constants/networkFilterModes';
 
 export const nodes = [
     {